Added JWT authentification logic and adjusted the token generation based of User id

+++ back_end/src/main/java/com/example/saastest/modules/benutzer/repository/BenutzerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BenutzerRepository extends JpaRepository<Benutzer, Integer> {
+public interface BenutzerRepository extends JpaRepository<Benutzer, Long> {
     Optional<Benutzer> findByEmail(String email);
 }
